+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 
 ## [Unreleased]
 
+### Added
+ - `ValueStringBuilder` constructor can take initial buffer instead of creating it itself.
+
 ## [1.1.0] - 2022-04-16
 
 ### Added